[TS-377] Refactor: habit item style file move
